requirements.txt
requests
MyQR
Retry
watchdog
python-telegram-bot==11.1.0